fix: move createTempCert to integration test file to fix unused lint error
The helper is only used by integration-tagged tests. Moving it to the
integration test file avoids the golangci-lint unused function warning
since lint runs without the integration build tag.
